moving tests to match repository structure
